9_vector_databases/4_qdrant/requirements.txt
langchain-openai
langchain_qdrant
langchain
python-dotenv
qdrant-client